backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0
Pillow==10.4.0
python-decouple==3.8
django-filter==24.3
